EpicBrothers/requirements.txt
pillow
django-crispy-forms
django-rest-framework
social-auth-app-django
django-pwa
